Refactor configuration files for neovim
The files are being refactored so that only two are used instead of six,
and the code is less repetitive.

The git editor configuration now has more in common with the config for
the general editor, and the general editor configuration sources the one
for git editor and adds only shortcuts for NERDTree and commands
responsible for displaying it by default.