Replace old configuration of i3blocks
The configuration is being replaced with a new one using a comination of
blocklets provided by anachron/i3blocks and vivien/i3blocks-contrib
projects.

No git submodules of these projects are used - instead, a selection of
only the blocks used in configuration has been copied into a common
blocklet directiory. This way, it's easier to refer to the blocklet
scripts in the configuration. The downside is that now there is no way
to automatically download the updates or to update the blocklets with a
simple command, but the use of these blocklets is only temporary - I'd
like to replace them with my custom scripts written in python.